+++ src/main/java/com/elmakers/mine/bukkit/utility/CompatibilityUtils.java
@@ -1228,4 +1228,14 @@ public class CompatibilityUtils extends NMSUtils {
         }
         return true;
     }
+
+    public static void applyItemData(ItemStack item, Block block) {
+        try {
+            Object entityDataTag = getNode(item, "BlockEntityTag");
+            if (entityDataTag == null) return;
+            NMSUtils.setTileEntityData(block.getLocation(), entityDataTag);
+        } catch (Exception ex) {
+            ex.printStackTrace();
+        }
+    }
 }