+++ src/main/java/com/elmakers/mine/bukkit/magic/MagicController.java
@@ -263,6 +263,9 @@ public class MagicController implements MageController {
             if (apiMage instanceof com.elmakers.mine.bukkit.magic.Mage) {
                 com.elmakers.mine.bukkit.magic.Mage mage = (com.elmakers.mine.bukkit.magic.Mage) apiMage;
 
+                // In case of rapid relog, this mage may have been marked for removal already
+                mage.setUnloading(false);
+                
                 // Re-set mage properties
                 mage.setName(mageName);
                 mage.setCommandSender(commandSender);
@@ -2743,7 +2746,24 @@ public class MagicController implements MageController {
     }
 
     public void playerQuit(Mage mage) {
-        playerQuit(mage, null);
+        // Delay logout one tick to avoid issues with plugins that kill
+        // players on logout (CombatTagPlus, etc)
+        // Don't delay on shutdown, though.
+        if (initialized && mage instanceof com.elmakers.mine.bukkit.magic.Mage) {
+            final com.elmakers.mine.bukkit.magic.Mage quitMage = (com.elmakers.mine.bukkit.magic.Mage)mage;
+            quitMage.setUnloading(true);
+            plugin.getServer().getScheduler().runTaskLater(plugin, new Runnable() {
+                @Override
+                public void run() {
+                    // Just in case the player relogged in that one tick..
+                    if (quitMage.isUnloading()) {
+                        playerQuit(quitMage, null);
+                    }
+                }
+            },1 );
+        } else {
+            playerQuit(mage, null);
+        }
     }
 
     protected void mageQuit(Mage mage) {