+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Add "Magic.use.<wand>" permission list, for controlling individual wand permissions.
  - Mine/Chop will no longer make drops out of temporary blocks
  - Second page of book can be used to set name/lore in /mitem skull command
+ - Add Magic.spawners permission, to allow non-op'd players to place custom mob spawners
 
 ## 5.5.5