+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add Magic.commands.cast.parameters permission, in case you really really want to give your players /cast permission.
  - Add Magic.spawners permission, to allow non-op'd players to place custom mob spawners
  - Converted tweak/upgrade icons to RP-based icons
+ - Several spells are now optional, players don't need to acquire them to rank up
 
 ## 6.1.1