+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - Second page of book can be used to set name/lore in /mitem skull command
  - Add Magic.spawners permission, to allow non-op'd players to place custom mob spawners
  - Add Magic.commands.cast.parameters permission, in case you really really want to give your players /cast permission.
+ - Add "/mtrait hat" command, for putting spells, wands and hats on an NPC's head
 
 ## 5.5.5