+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - Several spells are now optional, players don't need to acquire them to rank up
  - Add "/mtrait hat" command, for putting spells, wands and hats on an NPC's head (See: http://imgur.com/a/FQLnp)
  - Add support for an "override" entry in spells.yml, for settings that should effect ALL spells
+ - Fix compatibility issues with keep-on-death wands and CombatTagPlus kill-on-quit
 
 ## 6.1.1