+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 6.2
+## 6.1.2
 
  - Add "Soul Wand" system (currently work-in-progress)
  - Fix cyclemode upgrade